Delete event working and refreshing

+++ smvcp1/src/main/java/pep/mendez/smvcp1/spring/model/service/EventService.java
@@ -2,9 +2,6 @@ package pep.mendez.smvcp1.spring.model.service;
 
 import java.util.Collection;
 
-
-
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,15 +12,15 @@ import pep.mendez.smvcp1.spring.model.repository.EventRepository;
 @Service
 @Transactional
 public class EventService {
-	
+
 	@Autowired
 	private EventRepository eventRepository;
-	
+
 	@Transactional(readOnly = true)
 	public Collection<EventEntity> findAll() {
 		return eventRepository.findAll();
 	}
-	
+
 	@Transactional(readOnly = false)
 	public EventEntity save(EventEntity eventEntity) {
 		return eventRepository.save(eventEntity);
@@ -31,9 +28,11 @@ public class EventService {
 
 	@Transactional(readOnly = false)
 	public void delete(Long id) {
-		eventRepository.delete(id);
+		if (eventRepository.exists(id)) {
+			eventRepository.delete(id);
+		}
 	}
-	
+
 	@Transactional(readOnly = true)
 	public EventEntity find(Long id) {
 		return eventRepository.findOne(id);
